# MarketPlaceWebServiceExpress/docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: SML Staff Marketplace API
  version: "1.0.0"
  description: |
    Express.js port of the original Java/JAX-RS MarketPlaceWebService.

    All endpoints are under `/service/v1` (included in the server URL below).

    **Auth note**: Failed login returns HTTP 200 with `data: []` — not HTTP 401.

    **Image endpoints**: Return binary `image/png`. `/imagesguid` never returns 404 — returns empty buffer instead.
servers:
  - url: http://localhost:47300/service/v1
    description: Local development
...
security:
  - GUIDHeader: []
components:
  securitySchemes:
    GUIDHeader:
      type: apiKey
      in: header
      name: GUID
      description: Session identifier required by most endpoints
  schemas:
    ErrorResponse:
      type: object
      properties:
        ERROR:
          type: string
          example: "database connection error"
...
    CustomerRecord:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        address:
          type: string
        telephone:
          type: string
        tax_id:
          type: string
    EmployeeRecord:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
    CustomerCRMRecord:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        address:
          type: string
        telephone:
          type: string
        logistic_area:
          type: string
        gps:
          type: string
          description: GPS coords stored in ar_customer.website field
        group_main:
          type: string
    CustomerCRMPagination:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        current_page:
          type: integer
        total_page:
          type: integer
...
paths:
...
  /getCustomerList:
    get:
      tags: [Customer]
      summary: รายชื่อลูกค้า (max 50)
      description: Dropdown/search — hard limit 50
      parameters:
        - { name: search, in: query, schema: { type: string }, description: "ILIKE ค้นหา code หรือ name_1" }
      responses:
        '200':
          description: รายชื่อลูกค้า
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/CustomerRecord' }
        '400':
          description: Error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /getEmployeeList:
    get:
      tags: [Customer]
      summary: รายชื่อพนักงาน (max 50)
      description: รายชื่อจาก erp_user; ใช้ selectable_only เพื่อซ่อนรายการที่ mobile_user=1
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: code, in: query, schema: { type: string } }
        - name: selectable_only
          in: query
          description: When true or 1, exclude erp_user rows where mobile_user=1.
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: รายชื่อพนักงาน
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/EmployeeRecord' }
        '400':
          description: Error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /getCustomerCRM:
    get:
      tags: [Customer]
      summary: รายชื่อลูกค้า CRM (paginated)
      description: พร้อม logistic_area, GPS (จาก website field), group_main
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: limit,  in: query, schema: { type: integer, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: รายชื่อลูกค้า + pagination
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/CustomerCRMRecord' }
                  pagination: { $ref: '#/components/schemas/CustomerCRMPagination' }
        '400':
          description: Error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /getEmployeeCRM:
    get:
      tags: [Customer]
      summary: รายชื่อพนักงาน (paginated)
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: limit,  in: query, schema: { type: integer, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: รายชื่อพนักงาน
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/EmployeeRecord' }
        '400':
          description: Error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # ─── FINANCIAL ───────────────────────────────────────────────────────────────
...
  /getDimensionList:
    get:
      tags: [Customer]
      summary: รายการ Dimension (ar_dimension)
      parameters:
        - { name: search, in: query, schema: { type: string } }
      responses:
        '200':
          description: รายการ dimension
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        code:   { type: string }
                        name_1: { type: string }
  /getLogisticAreaList:
    get:
      tags: [Customer]
      summary: รายการเขตโลจิสติกส์ (ar_logistic_area)
      parameters:
        - { name: search, in: query, schema: { type: string } }
      responses:
        '200':
          description: รายการเขต
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        code:   { type: string }
                        name_1: { type: string }
  /getCustomerManageList:
    get:
      tags: [Customer]
      summary: รายชื่อลูกค้าสำหรับจัดการ (paginated, ข้อมูลครบ)
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: limit,  in: query, schema: { type: integer, default: 30 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: รายชื่อลูกค้า
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/CustomerCRMRecord' }
                  pagination: { $ref: '#/components/schemas/CustomerCRMPagination' }
  /getNextCustomerCode:
    get:
      tags: [Customer]
      summary: สร้างรหัสลูกค้าถัดไป
      description: Generate รหัสแบบ OR-XXXXX โดย auto-increment จาก ar_customer
      responses:
        '200':
          description: รหัสลูกค้าถัดไป
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  code:    { type: string, example: "OR-00042" }
        '409':
          description: ไม่สามารถสร้างรหัสที่ไม่ซ้ำได้
  /getCustomerDetail:
    get:
      tags: [Customer]
      summary: รายละเอียดลูกค้า (ข้อมูลครบ รวม credit, logistic, sale)
      parameters:
        - { name: code, in: query, required: true, schema: { type: string }, example: "C00001" }
      responses:
        '200':
          description: รายละเอียดลูกค้า หรือ data:null ถ้าไม่พบ
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    nullable: true
                    allOf:
                      - $ref: '#/components/schemas/CustomerCRMRecord'
        '400':
          description: ไม่ระบุรหัสลูกค้า
  /createCustomer:
    post:
      tags: [Customer]
      summary: สร้างลูกค้าใหม่
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, name_1, password]
              properties:
                code:        { type: string, example: "OR-00042" }
                name_1:      { type: string, example: "บริษัท ใหม่ จำกัด" }
                address:     { type: string }
                telephone:   { type: string }
                email:       { type: string }
                website:     { type: string, description: "ใช้เก็บ GPS coordinates" }
                price_level: { type: integer, default: 0 }
                tax_id:      { type: string }
                group_main:  { type: string }
                logistic_area: { type: string }
                sale_code:   { type: string }
                credit_day:  { type: integer }
                password:    { type: string, description: "รหัสผ่านลูกค้า (เก็บใน ar_customer.fax)" }
      responses:
        '200':
          description: สร้างสำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { type: object }
        '400':
          description: ข้อมูลไม่ครบ
        '409':
          description: รหัสลูกค้านี้มีอยู่แล้ว
  /updateCustomer:
    post:
      tags: [Customer]
      summary: แก้ไขข้อมูลลูกค้า (สำหรับ admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, name_1]
              properties:
                code:     { type: string }
                name_1:   { type: string }
                address:  { type: string }
                telephone: { type: string }
                email:    { type: string }
                website:  { type: string }
                price_level: { type: integer }
                tax_id:   { type: string }
                group_main: { type: string }
                logistic_area: { type: string }
                sale_code: { type: string }
                credit_day: { type: integer }
                password: { type: string }
      responses:
        '200':
          description: แก้ไขสำเร็จ
        '400':
          description: ข้อมูลไม่ครบ
        '404':
          description: ไม่พบรหัสลูกค้า
  /updateCustomerProfile:
    post:
      tags: [Customer]
      summary: แก้ไขข้อมูลส่วนตัวลูกค้า (สำหรับลูกค้าแก้ไขเอง)
      description: แก้ได้เฉพาะ name_1, address, telephone, email (และ password ถ้าส่งมา)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, name_1]
              properties:
                code:      { type: string }
                name_1:    { type: string }
                address:   { type: string }
                telephone: { type: string }
                email:     { type: string }
                password:  { type: string, description: "ถ้าไม่ส่งมา → ไม่เปลี่ยน password" }
      responses:
        '200':
          description: แก้ไขสำเร็จ
        '400':
          description: ข้อมูลไม่ครบ
        '404':
          description: ไม่พบรหัสลูกค้า
  /deleteCustomer:
    post:
      tags: [Customer]
      summary: ลบลูกค้า
      description: |
        ตรวจสอบก่อนว่าลูกค้าถูกใช้ใน ic_trans หรือไม่ — ถ้าใช้แล้วจะไม่ลบ

        ลบ ar_item_by_customer, ar_customer_detail, ar_customer ตามลำดับ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string, example: "OR-00042" }
      responses:
        '200':
          description: ลบสำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
        '400':
          description: ไม่ระบุรหัสลูกค้า
        '404':
          description: ไม่พบรหัสลูกค้า
        '409':
          description: ลบไม่ได้เพราะถูกใช้ในเอกสารแล้ว
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:    { type: boolean }
                  message:    { type: string }
                  used_count: { type: integer }

  # ─── POS (เพิ่มเติม) ─────────────────────────────────────────────────────────
...